Make new notification form using Remco form
